Simplify CustomCSS UI
